+++ silk-react-components/CHANGELOG.md
@@ -1,1 +1,5 @@
 ## [Unreleased]
+
+### Add
+
+- unit tests for silk-react-components